ロジックをhook.tsxに切り出し
